refactor: move all js into utils
