+++ utils/background.js
